Basket show value

+++ app/src/main/java/com/example/coffeeappday1a/data/model/CoffeeDrink.kt
@@ -3,11 +3,11 @@ package com.example.coffeeappday1a.data.model
 import java.io.Serializable
 
 data class CoffeeDrink(
-    val id: Int,
+    val id: Long,
     val name: String,
     val type: DrinkType,
     val size: Size,
     val volumeMl: Int,
     val price: Double,
-    var quantity: Int = 1   // 👈 добавили количество
+    var quantity: Int = 1
 ) : Serializable
